:cherry_blossom: pkg,sql

+++ pkg/response/index.go
@@ -1,10 +1,10 @@
 package response
 
 type ArticleIndexList struct {
-	ID           int    `json:"id"`
+	ID           int64  `json:"id"`
 	Title        string `json:"title"`
 	Description  string `json:"description"`
 	IsTop        int8   `json:"is_top"`
-	CategoryID   int    `json:"category_id"`
+	CategoryID   int64  `json:"category_id"`
 	CategoryName string `json:"categoryName"`
 }
